fix: Warning: comparison of integers of different signs
The warning could be ignored as any negative value for `ret` is handled
before. To fix the warning just cast `ret` to `size_t`.

+++ src/atecc508a.c
@@ -191,7 +191,7 @@ static int atecc508a_read(const struct atecc508a *const priv,
 			retry++;
 			if (retry >= MAX_RETRY)
 				return -1;
-		} else if (ret != sizeof(struct atecc508a_recv_data) - num_read) {
+		} else if ((size_t)ret != sizeof(struct atecc508a_recv_data) - num_read) {
 			num_read += ret;
 			continue;
 		} else {
